+++ src/main/java/com/evdealer/ev_dealer_management/car/model/dto/color/ColorPostDto.java
@@ -1,9 +1,11 @@
 package com.evdealer.ev_dealer_management.car.model.dto.color;
 
 
+import java.math.BigDecimal;
+
 public record ColorPostDto (
     String colorName,
     String colorHex,
-    Float extraCost
+    BigDecimal extraCost
 ) {
 }
