+++ src/main/java/com/evdealer/ev_dealer_management/car/model/dto/performance/PerformancePostDto.java
@@ -5,10 +5,10 @@ import com.evdealer.ev_dealer_management.car.model.enumeration.ChemistryType;
 public record PerformancePostDto(
         Long batteryId,
         Long motorId,
-        Double rangeMiles,
-        Double accelerationSec,
-        Double topSpeedMph,
-        Double towingLbs
+        Float rangeMiles,
+        Float accelerationSec,
+        Float topSpeedMph,
+        Float towingLbs
 ) {
 }
